// apps/frontend/src/api/generated/models/videoVisibility.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Route Streaming API.
 * Route Streaming Service Backend API.
 * OpenAPI spec version: 1.0.
 */

export type VideoVisibility =
  (typeof VideoVisibility)[keyof typeof VideoVisibility]

export const VideoVisibility = {
  Private: 'private',
  Public: 'public',
} as const